deploy, ci: add deployment manifests
This commit adds some `base` deployment manifests to set up the
provisioner, and a `ci` overlay which includes a couple of objects that
consume storage 'provisioned' by `cosi-provisioner-sample`.

It also extends the GitHub Actions pipeline to deploy a Kind cluster in
which the COSI CRDs, controller and CSI plugin are deployed, and the
`ci` overlay of this project. We then validate the expected storage
access configuration files become available in a test Pod.