cosi-driver-sample: rework datamodel
Instead of keeping a set of accounts 'within' a bucket structure, which
is likely quite unlike a 'real' storage system, keep buckets and
accounts as separate top-level objects.

This also allows for `ProvisionerRevokeBucketAccess` to (still) function
when the bucket was deleted before, which is a prerequisite for some
COSI workflows.

Discussed during the community meeting of May 10th, 2021.